+++ packages/nesd/lib/ui/emulator/input/action_handler.g.dart
@@ -89,4 +89,4 @@ final class ActionHandlerProvider
   }
 }
 
-String _$actionHandlerHash() => r'a80524c0b93f5c990f4ad1041dd41c553d41e3b0';
+String _$actionHandlerHash() => r'fb39d09bd1803e465e641ec12dd4084c99cfa329';