+++ packages/nesd/lib/ui/main_menu/recent_rom_list.dart
@@ -25,7 +25,7 @@ class RecentRomList extends HookConsumerWidget {
       settingsControllerProvider.select((settings) => settings.recentRoms),
     );
 
-    final route = ref.watch(routerObserverProvider);
+    final route = ref.watch(currentRouteProvider);
 
     if (recentRoms.isEmpty) {
       return Center(
@@ -69,13 +69,11 @@ class RecentRomList extends HookConsumerWidget {
           for (final romTileData in roms)
             RomTile(
               onPressed: () async {
-                final success = await controller.loadRom(
+                final started = await controller.startRom(
                   romTileData.romInfo.file,
                 );
 
-                if (success || !context.mounted) {
-                  ref.read(routerProvider).navigate(const EmulatorRoute());
-
+                if (started || !context.mounted) {
                   return;
                 }