+++ packages/nesd/test/ui/emulator/nes_controller_isolate_test.dart
@@ -7,6 +7,7 @@ import 'package:nesd/ui/emulator/nes_controller.dart';
 import 'package:nesd/ui/emulator/rom_manager.dart';
 import 'package:nesd/ui/file_picker/file_system/filesystem.dart';
 import 'package:nesd/ui/file_picker/file_system/filesystem_file.dart';
+import 'package:nesd/ui/router/router.dart';
 import 'package:nesd/ui/settings/settings.dart';
 import 'package:nesd/ui/toast/toaster.dart';
 
@@ -83,6 +84,7 @@ void main() {
     final controller = NesController(
       nesState: container.read(nesStateProvider.notifier),
       spawner: spawner,
+      router: Router(),
       settingsController: settings,
       toaster: _MockToaster(),
       romManager: romManager,